Version 2
In this version, I added the All subjects page as well as the quiz page. I added the All subjects page in preparation for the next commit where I will add the list of new pages for the subjects the end user will visit.

I also added a number of styling changes to make the site look better, in particular to ensure that resizing the page will not break the design of the page any longer.

Commenting of HTML and CSS pages still to come later down the line
Still need to add .js files for the quiz.