Version 3
Added all of the subject pages. In addition to this, all pages have been linked together so that all buttons now work.

Fixed a bug where the page the user was on would not appear as the one that was active in the sidebar (e.g if I was on a subject page, the All subjects page that I came from would not be lit up, but instead the quiz would be, vice versa etc.)

Added a global.css file to make my CSS stylesheet linking more efficient (get rid of every single individual CSS stylesheet for every webpage.)

Still need to comment code and add JavaScript.