Version 4
Added the tips and tricks page that I intended to add since the beginning.

Took out all the individual CSS stylesheets for each individual  subject page and linked the global.css page to all subject pages.

Added in the quiz JavaScript! The quiz is perfectly functional!
Also added JavaScript for the pop-up to show the results of the quiz.

Commenting was completed for the JavaScript file. Will complete later for the HTML/CSS files.